fix(tools/setup): throw if directory exists.

+++ tools/setup.ts
@@ -1,13 +1,17 @@
 import { parseArgs } from "jsr:@std/cli/parse-args";
 import { join, resolve, relative, dirname } from '@std/path'
-import { walk, ensureDir } from '@std/fs'
+import { walk, ensureDir, exists } from '@std/fs'
 import { parse, format, type SemVer } from '@std/semver'
 import { createRegExp, exactly, maybe, wordChar, anyOf, digit } from 'magic-regexp'
 import { capitalize, snakeCase } from 'cosmokit'
+import { Ask } from "jsr:@sallai/ask";
 import z from 'schemastery'
+import mkdir = Deno.mkdir;
 
 export type PackType = 'package' | 'plugin'
 export type Scope = 'p' | 'plug'
+
+const ask = new Ask();
 
 function toScope(typ: PackType): Scope | null {
   if (typ === 'package') return 'p'
@@ -37,7 +41,7 @@ const VersionSchema = z.object({
 const schema = z.object({
   webui: z.boolean().default(false).experimental(),
   // deno-lint-ignore no-explicit-any
-  name: z.string().default(null as any),
+  name: z.string().pattern(regexp).default(null as any),
   package: z.boolean().default(false),
   version: z.union([
     VersionSchema.default('1.0.0'),
@@ -49,9 +53,12 @@ const args = schema(parseArgs(Deno.args))
 let type: PackType = args.package ? 'package' : 'plugin'
 if (args.webui) throw new Error("webui template is not supported yet")
 if (!args.name) {
-  const prompted = prompt(`${capitalize(type)} Name >`)
-  if (!prompted) throw new Error("plugin name is not provided")
-  args.name = prompted
+  const { name } = await ask.input({
+    name: "name",
+    message: `${capitalize(type)} Name ›`,
+  } as const);
+  if (!name) throw new Error("plugin name is not provided")
+  args.name = z.string().pattern(regexp)(name)
 }
 
 const nameSchema = z.object({
@@ -63,13 +70,11 @@ const matched = args.name.match(regexp)
 // deno-lint-ignore no-explicit-any
 const { scope, name } = nameSchema(<any>(matched!.groups))
 if (toScope(type) != scope) {
-  console.warn(`expected a scope of ${toScope(type)}, got ${scope}`)
+  console.warn(`%cwarning%c: expected a scope of ${toScope(type)}, got ${scope}`, 'color: yellow')
   type = fromScope(scope)!
 }
 
 console.log(`Creating ${capitalize(fromScope(scope)!)} @${scope}/${name}`)
-
-console.log(args)
 
 function interpolate(content: string) {
   return content
@@ -80,13 +85,15 @@ function interpolate(content: string) {
 
 const template_dir = resolve(import.meta.dirname!, '__template__')
 const target_dir = join(resolve(Deno.cwd(), type + 's'), snakeCase(name))
+if (await exists(target_dir)) throw new Error(relative(Deno.cwd(), target_dir) + " already exists")
 for await (const entry of walk(template_dir)) {
   // if (entry.name === '__template__') continue;
   const target_path = join(target_dir, relative(template_dir, entry.path))
-  if (entry.isDirectory) await ensureDir(target_path)
-  await ensureDir(dirname(target_path))
-  await Deno.writeTextFile(
-    target_path,
-    interpolate(await Deno.readTextFile(entry.path))
+  if (entry.isDirectory) await mkdir(target_path, { recursive: true })
+  else
+    await Deno.writeTextFile(
+      target_path,
+      interpolate(await Deno.readTextFile(entry.path)),
+      { createNew: true }
   )
 }
